+++ src/controllers/__tests__/signinController.spec.ts
@@ -1,7 +1,7 @@
 import casual from 'casual';
 import { Request, Response } from 'express';
 import { Cache } from "../../datasources/cache";
-import { generateTokens, setTokenCookie } from '../../services/tokenService';
+import { generateAuthTokens, setTokenCookie } from '../../services/tokenService';
 import { generalConfig } from '../../config/generalConfig';
 import { signinController } from '../signinController';
 import * as UserModel from '../../models/User';
@@ -19,6 +19,7 @@ const mockedUser: UserModel.User = {
   affiliationId: casual.url,
   role: UserModel.UserRole.RESEARCHER,
   password: casual.uuid,
+  acceptedTerms: true,
   created: new Date().toISOString(),
   errors: [],
 
@@ -64,14 +65,14 @@ describe('signinController', () => {
 
   it('should sign the user in and set the access and refresh tokens successfully', async () => {
     jest.spyOn(mockUser, 'login').mockResolvedValueOnce(mockedUser);
-    (generateTokens as jest.Mock).mockResolvedValue({
+    (generateAuthTokens as jest.Mock).mockResolvedValue({
       accessToken: 'new-access-token',
       refreshToken: 'new-refresh-token'
     });
 
     await signinController(mockRequest as Request, mockResponse as Response);
 
-    expect(generateTokens).toHaveBeenCalledWith(mockCache, mockedUser);
+    expect(generateAuthTokens).toHaveBeenCalledWith(mockCache, mockedUser);
     expect(setTokenCookie).toHaveBeenCalledWith(mockResponse, 'dmspt', 'new-access-token', generalConfig.jwtTTL);
     expect(setTokenCookie).toHaveBeenCalledWith(mockResponse, 'dmspr', 'new-refresh-token', generalConfig.jwtRefreshTTL);
     expect(mockResponse.status).toHaveBeenCalledWith(200);
@@ -88,7 +89,7 @@ describe('signinController', () => {
 
   it('should return 500 if unable to generate tokens', async () => {
     jest.spyOn(mockUser, 'login').mockResolvedValueOnce(mockedUser);
-    (generateTokens as jest.Mock).mockResolvedValue({});
+    (generateAuthTokens as jest.Mock).mockResolvedValue({});
 
     await signinController(mockRequest as Request, mockResponse as Response);
 
@@ -99,7 +100,7 @@ describe('signinController', () => {
   it('should return 500 if an unexpected error occurs', async () => {
     jest.spyOn(mockUser, 'login').mockResolvedValueOnce(mockedUser);
     const mockError = new Error('Unexpected error');
-    (generateTokens as jest.Mock).mockRejectedValue(mockError);
+    (generateAuthTokens as jest.Mock).mockRejectedValue(mockError);
 
     await signinController(mockRequest as Request, mockResponse as Response);