+++ src/index.ts
@@ -2,6 +2,7 @@ import { ApolloServer } from '@apollo/server';
 import express from 'express';
 import http from 'http';
 import cors from 'cors';
+import cookieParser from 'cookie-parser';
 import { logger } from './logger';
 import { serverConfig } from './config';
 import { healthcheck } from './controllers/healthcheck';
@@ -9,6 +10,7 @@ import { attachApolloServer } from './middleware/express';
 import router from './router';
 import { MySQLDataSource } from './datasources/mySQLDataSource';
 import { Cache } from './datasources/cache';
+import { csrfMiddleware } from './middleware/csrf';
 
 // TODO: Make this configurable and pass in as ENV variable
 const PORT = 4000;
@@ -30,9 +32,11 @@ const startServer = async () => {
   // Healthcheck endpoint (declare this BEFORE CORS definition due to AWS ALB limitations)
   app.get('/up', (_request, response) => healthcheck(apolloServer, response, logger));
 
-  // Express middleware
+  // Express middleware for all requests (besides the healthcheck above)
   app.use(
+    cookieParser(),
     cors(),
+    csrfMiddleware,
     express.urlencoded({ extended: false }),
     express.json({ limit: '50mb' }),
   )